// src/bcp-client/src/app/api/model/updateInscriptionRequest.ts
/**
 * BilliardClubParisien.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InscriptionCommand } from './inscriptionCommand';


export interface UpdateInscriptionRequest { 
    data: InscriptionCommand;
}